# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: AI Code Reviewer browser API
  version: 0.1.0
  description: |
    Browser-facing HTTP surface of the AI Code Reviewer: `portal-api` (runs, repositories,
    pull requests, live updates) and `auth-api` (GitHub App user authorization), as decided
    in api#20 (D4, D9, D10, D11). The GitHub webhook receiver (`/webhooks/github`) and
    `/healthcheck` are not part of the browser surface and are not described here.

    **Origin and CORS.** The UI and the API share one origin: the UI host serves `/api/*` and
    its proxy (the edge proxy or the UI's nginx; the Vite dev server locally) forwards those
    paths to the API, so there is no CORS (#35). A separate API host remains only for
    `/healthcheck` and GitHub webhooks. Until the service split (#16, later), the `auth-api`,
    `portal-api` and `webhook-api` routes live in the single FastAPI app `app/main.py`.

    **Wire format.** JSON bodies use camelCase keys. Timestamps are RFC 3339 in UTC with a `Z`
    suffix. IDs are UUIDs unless a field says it carries a GitHub numeric ID.

    **Authentication.** `auth-api` issues a 15-minute access JWT, sent as
    `Authorization: Bearer <token>` on every other operation (global `bearerAuth`), and a
    rotating `refresh_token` cookie (30 days, `HttpOnly; Secure; SameSite=Strict;
    Path=/api/auth`). The SSE stream is read with `fetch` so it carries the same header.
...
    in #11.

...
    request validation failures are FastAPI's `HTTPValidationError`.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/github/callback:
    post:
      operationId: exchangeGithubCode
      tags: [auth]
      summary: Exchange a GitHub App user authorization code for a session
      description: >-
        The SPA generates and checks `state` itself (sessionStorage) and posts only the code
        GitHub returned to `/auth/callback`. Sets the refresh cookie.
      x-service: auth-api
      x-issue: "#11"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCallbackRequest"
      responses:
        "503":
          description: Authentication is not configured.
        "200":
          description: A new session.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SetRefreshCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "502":
          description: GitHub authentication is unavailable.
        "400":
          description: The code is invalid, expired or already used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Rotate the refresh cookie and issue a new access token
      x-service: auth-api
      x-issue: "#11"
      security:
        - refreshCookie: []
      responses:
        "503":
          description: Authentication is not configured.
        "200":
          description: A new access token; the refresh cookie is rotated.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SetRefreshCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: The refresh cookie is missing, expired or revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the signed-in user and the workspaces they can access
      x-service: auth-api
      x-issue: "#11"
      responses:
        "503":
          description: Authentication is not configured.
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Revoke the refresh token and clear the cookie
      description: >-
        Idempotent; revokes the refresh token from the `refresh_token` cookie when one is
        sent, and succeeds without it as well.
      x-service: auth-api
      x-issue: "#11"
      security: []
      responses:
        "503":
          description: Authentication is not configured.
        "204":
          description: Signed out.
          headers:
            Set-Cookie:
              description: Expires the `refresh_token` cookie (`Max-Age=0; Path=/api/auth`).
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT issued by auth-api, valid for 15 minutes.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >-
        Rotating refresh token, 30 days, `HttpOnly; Secure; SameSite=Strict; Path=/api/auth`.
...
  headers:
    SetRefreshCookie:
      description: >-
        `refresh_token=<token>; Max-Age=2592000; Path=/api/auth; HttpOnly; Secure;
        SameSite=Strict`.
      schema:
        type: string
  responses:
    Unauthorized:
      description: The access token is missing, invalid or expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationFailed:
      description: A path, query or body parameter failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
    ErrorResponse:
      type: object
      description: Body of every `HTTPException` error.
      required: [detail]
      properties:
        detail:
          type: string
      additionalProperties: false
    HTTPValidationError:
      type: object
      description: FastAPI request validation error.
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
        input: {}
        ctx:
          type: object
...
    User:
      type: object
      required: [id, login, name, avatarUrl]
      properties:
        id:
          type: integer
          description: GitHub user ID.
        login:
          type: string
          minLength: 1
        name:
          type: [string, "null"]
        avatarUrl:
          type: [string, "null"]
          format: uri
      additionalProperties: false
    Workspace:
      type: object
      required: [id, name, installationId]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        installationId:
          type: integer
          description: GitHub App installation ID.
      additionalProperties: false
    Me:
      type: object
      description: >-
        User plus the workspaces the user can access through GitHub App installations.
        Repeats every User property (tests/test_openapi_contract.py checks that they stay
        identical).
      required: [id, login, name, avatarUrl, workspaces]
      properties:
        id:
          type: integer
          description: GitHub user ID.
        login:
          type: string
          minLength: 1
        name:
          type: [string, "null"]
        avatarUrl:
          type: [string, "null"]
          format: uri
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
      additionalProperties: false
    AuthCallbackRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          description: The `code` query parameter GitHub appended to `/auth/callback`.
      additionalProperties: false
    AuthSession:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          minLength: 1
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          minimum: 1
          description: Access token lifetime in seconds (900).
        user:
          $ref: "#/components/schemas/User"
      additionalProperties: false
...